// apps/job-service/src/modules/applicator-workers/applicator-workers-types.ts
import { WorkerType } from '@prisma/client';
import { Decimal } from '@prisma/client/runtime/library';

interface ApplicatorWorker {
	title: WorkerType;
	firstName: string;
	lastName: string;
	email: string;
	phoneNumber?: string;
	businessName?: string;
	address1?: string;
	address2?: string;
	state: string;
	county: string;
	township: string;
	zipCode: string;
    pilotLicenseNumber?:string;
    businessLicenseNumber?:string;
    planeOrUnitNumber?:string;
    perAcrePricing:Decimal;
    percentageFee: Decimal;       
    dollarPerAcre: Decimal;
    autoAcceptJobs:boolean        
    canViewPricingDetails:boolean 
    code:string
    lastLogin: Date   
}

export {ApplicatorWorker}